Quiet a compiler warning.

+++ sender.c
@@ -47,8 +47,8 @@ void read_sum_head(int f, struct sum_struct *sum)
 	} else {
 		sum->s2length = read_int(f);
 		if (sum->s2length > MD4_SUM_LENGTH) {
-			rprintf(FERROR, "Invalid checksum length %d\n",
-			    sum->s2length);
+			rprintf(FERROR, "Invalid checksum length %ld\n",
+			    (long)sum->s2length);
 			exit_cleanup(RERR_PROTOCOL);
 		}
 	}
